fix(cli): the published binary reads EVA_LOG, and the renderer is named where a reader looks it up
EVA_LOG was read by the workspace entry alone, so a person running the published binary got none of it. It is read once, where every door into Eva goes through.
