+++ apps/cli/src/index.ts
@@ -5,7 +5,7 @@ import type { Claim } from "@missingstudio/eva-schema"
 import { grantTrust, isTrusted, revokeTrust, type Overlays } from "@missingstudio/eva-kernel"
 import { checkRules, sayFault, unreachableIn } from "@missingstudio/eva-tool-policy"
 import { refusal } from "@missingstudio/eva-web"
-import { Effect, Exit, Scope } from "effect"
+import { Effect, Exit, Logger, LogLevel, References, Scope, Tracer } from "effect"
 import { parseArgv, showHelp } from "./argv.js"
 import { runAttach } from "./attach.js"
 import { attaching, besideTerminal, BUILD, serving } from "./plugins.js"
@@ -307,5 +307,58 @@ export const main = Effect.fn("cli.main")(function* (world: World, build: Build 
   }
 })
 
-export const run = (args?: readonly string[]): Promise<number> =>
-  Effect.runPromise(main(fromProcess(args)))
+/**
+ * The level `EVA_LOG` names, or nothing when it names none. The names are
+ * Effect's own, in any case: all, trace, debug, info, warn, error, fatal,
+ * none. A value that names no level leaves the run as it was.
+ */
+const levelOf = (asked: string | undefined): LogLevel.LogLevel | undefined =>
+  LogLevel.values.find((level) => level.toLowerCase() === asked?.toLowerCase())
+
+/**
+ * A tracer that says each span as it ends. Nearly every function in the tree
+ * is an `Effect.fn("…")`, so this is the call tree of a run, said innermost
+ * first as it unwinds. It writes to the error stream, because standard output
+ * is the answer.
+ */
+const saying = (err: World["err"]): Tracer.Tracer => {
+  class Said extends Tracer.NativeSpan {
+    override end(endTime: bigint, exit: Exit.Exit<unknown, unknown>): void {
+      super.end(endTime, exit)
+      const took = Number(endTime - this.startTime) / 1e6
+      err(`eva: ${this.name} ${took.toFixed(1)}ms${Exit.isFailure(exit) ? " failed" : ""}\n`)
+    }
+  }
+  return Tracer.make({ span: (options) => new Said(options) })
+}
+
+/**
+ * The run, with what `EVA_LOG` asked for over it. The level says what is
+ * logged and sends it to the error stream; a level of debug or finer opens
+ * the spans as well, because a span is a detail.
+ */
+const watched = <A, E>(
+  run: Effect.Effect<A, E>,
+  level: LogLevel.LogLevel,
+  err: World["err"],
+): Effect.Effect<A, E> => {
+  const logged = run.pipe(
+    Effect.provideService(References.MinimumLogLevel, level),
+    Effect.provideService(Logger.LogToStderr, true),
+  )
+  return LogLevel.isLessThanOrEqualTo(level, "Debug")
+    ? Effect.withTracer(logged, saying(err))
+    : logged
+}
+
+/**
+ * Every door into Eva. The published binary calls this, and so does the
+ * workspace entry, so `EVA_LOG` is read once here and reaches both.
+ */
+export const run = (args?: readonly string[]): Promise<number> => {
+  const world = fromProcess(args)
+  const level = levelOf(world.env["EVA_LOG"])
+  return Effect.runPromise(
+    level === undefined ? main(world) : watched(main(world), level, world.err),
+  )
+}
